tests/wpt/tests/html/canvas/offscreen/path-objects/2d.path.roundrect.badinput.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.path.roundrect.badinput</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.path.roundrect.badinput</h1>
<p class="desc">roundRect() throws or does not throw errors given the strange inputs.</p>


<script>
var t = async_test("roundRect() throws or does not throw errors given the strange inputs.");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.roundRect(0, 0, 100, 100, { foo: "bar" });  //=> DOMPointInit
  ctx.roundRect(0, 0, 100, 100, undefined);       //=> "missing" -> 0
  ctx.roundRect(0, 0, 100, 100, [[]]);            //=> « DOMPointInit »
  ctx.roundRect(0, 0, 100, 100, [[25]]);          //=> « DOMPointInit »
  ctx.roundRect(0, 0, 100, 100, [undefined]);     //=> « DOMPointInit »
  assert_throws_js(TypeError, function() { ctx.roundRect(0, 0, 100, 100, 0n); });
  assert_throws_js(TypeError, function() { ctx.roundRect(0, 0, 100, 100, { x: 0n }); });
  assert_throws_js(TypeError, function() { ctx.roundRect(0, 0, 100, 100, [{ x: 0n }]); });
  t.done();

});
</script>
